refactor: Split create move tests
This change refactors the create move end-to-end tests into separate
files for each journey.

We need to add further journeys so this will make it more manageable,
as well as offering us the possibility to run tests in parallel in
future to reduce overall test time.
